requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2     
attrs==22.2.0
certifi==2022.12.7       
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
fastapi==0.92.0
frozenlist==1.3.3        
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2        
multidict==6.0.4
openai==0.26.5
pydantic==1.10.5
python-dotenv==0.21.1    
requests==2.28.2
sniffio==1.3.0
starlette==0.25.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
yarl==1.8.2